// projects/ui/src/lib/components/po-skeleton/po-skeleton-animation.enum.ts
/**
 * @usedBy PoSkeletonComponent
 *
 * @description
 *
 * Define os tipos de animação do componente `po-skeleton`.
 */
export enum PoSkeletonAnimation {
  /** Sem animação */
  none = 'none',

  /** Animação de pulsação */
  pulse = 'pulse',

  /** Animação de brilho deslizante */
  shimmer = 'shimmer'
}