// projects/ui/src/lib/components/po-skeleton/po-skeleton-variant.enum.ts
/**
 * @usedBy PoSkeletonComponent
 *
 * @description
 *
 * Define as variantes visuais do componente `po-skeleton`.
 */
export enum PoSkeletonVariant {
  /** Variante para simular texto */
  text = 'text',

  /** Variante retangular */
  rect = 'rect',

  /** Variante circular */
  circle = 'circle'
}